refactor: ui 패키지 mvc 패턴으로 패키지 분리
